+++ CHANGES.rst
@@ -63,6 +63,7 @@ Development
 - Improve tooling and inline documentation
 - Add option ``main.fastboot`` for skipping LTE modem teardown
   for faster development iterations
+- Curate the garbage collector
 
 
 2019-06-22 0.5.1